change project to django
I thought about this for a bit and it occured to me that to make
things done fast we need to use Django here. Specifically we need
easy to use admin interface, couple of forms and REST API. With flask
it is harder to build these because they don't have common db engine.

I have more important things to do then building the piping around Flask.